+++ job/views.py
@@ -34,10 +34,16 @@ import traceback as tb
 from django.http import JsonResponse
 from reo.exceptions import UnexpectedError
 from job.models import Settings, PVInputs, ElectricStorageInputs, WindInputs, GeneratorInputs, ElectricLoadInputs,\
-    ElectricTariffInputs, ElectricUtilityInputs, SpaceHeatingLoadInputs, PVOutputs, ElectricStorageOutputs, WindOutputs, ExistingBoilerInputs,\
-    GeneratorOutputs, ElectricTariffOutputs, ElectricUtilityOutputs, ElectricLoadOutputs, ExistingBoilerOutputs, \
-    DomesticHotWaterLoadInputs, SiteInputs, SiteOutputs, APIMeta, UserProvidedMeta
-
+    ElectricTariffInputs, ElectricUtilityInputs, SpaceHeatingLoadInputs, PVOutputs, ElectricStorageOutputs,\
+    WindOutputs, ExistingBoilerInputs, GeneratorOutputs, ElectricTariffOutputs, ElectricUtilityOutputs,\
+    ElectricLoadOutputs, ExistingBoilerOutputs, DomesticHotWaterLoadInputs, SiteInputs, SiteOutputs, APIMeta,\
+    UserProvidedMeta, CHPInputs, CHPOutputs, CoolingLoadInputs, ExistingChillerInputs, ExistingChillerOutputs,\
+    CoolingLoadOutputs, HeatingLoadOutputs, HotThermalStorageInputs, HotThermalStorageOutputs,\
+    ColdThermalStorageInputs, ColdThermalStorageOutputs
+import os
+import requests
+import logging
+log = logging.getLogger(__name__)
 
 def make_error_resp(msg):
     resp = dict()
@@ -61,11 +67,16 @@ def help(request):
         d["ElectricStorage"] = ElectricStorageInputs.info_dict(ElectricStorageInputs)
         d["Wind"] = WindInputs.info_dict(WindInputs)
         d["Generator"] = GeneratorInputs.info_dict(GeneratorInputs)
+        d["CoolingLoad"] = CoolingLoadInputs.info_dict(CoolingLoadInputs)
+        d["ExistingChiller"] = ExistingChillerInputs.info_dict(ExistingChillerInputs)
         d["ExistingBoiler"] = ExistingBoilerInputs.info_dict(ExistingBoilerInputs)
         # d["Boiler"] = BoilerInputs.info_dict(BoilerInputs)
+        d["HotThermalStorage"] = HotThermalStorageInputs.info_dict(HotThermalStorageInputs)
+        d["ColdThermalStorage"] = ColdThermalStorageInputs.info_dict(ColdThermalStorageInputs)
         d["SpaceHeatingLoad"] = SpaceHeatingLoadInputs.info_dict(SpaceHeatingLoadInputs)
         d["DomesticHotWaterLoad"] = DomesticHotWaterLoadInputs.info_dict(DomesticHotWaterLoadInputs)
         d["Site"] = SiteInputs.info_dict(SiteInputs)
+        d["CHP"] = CHPInputs.info_dict(CHPInputs)
         return JsonResponse(d)
 
     except Exception as e:
@@ -99,8 +110,15 @@ def outputs(request):
         d["ElectricStorage"] = ElectricStorageOutputs.info_dict(ElectricStorageOutputs)
         d["Wind"] = WindOutputs.info_dict(WindOutputs)
         d["Generator"] = GeneratorOutputs.info_dict(GeneratorOutputs)
+        d["ExistingChiller"] = ExistingChillerOutputs.info_dict(ExistingChillerOutputs)
         d["ExistingBoiler"] = ExistingBoilerOutputs.info_dict(ExistingBoilerOutputs)
         # d["Boiler"] = BoilerOutputs.info_dict(BoilerOutputs)
+        d["HotThermalStorage"] = HotThermalStorageOutputs.info_dict(HotThermalStorageOutputs)
+        d["ColdThermalStorage"] = ColdThermalStorageOutputs.info_dict(ColdThermalStorageOutputs)
+        d["Site"] = SiteOutputs.info_dict(SiteOutputs)
+        d["HeatingLoad"] = HeatingLoadOutputs.info_dict(HeatingLoadOutputs)
+        d["CoolingLoad"] = CoolingLoadOutputs.info_dict(CoolingLoadOutputs)
+        d["CHP"] = CHPOutputs.info_dict(CHPOutputs)
         return JsonResponse(d)
 
     except Exception as e:
@@ -186,16 +204,31 @@ def results(request, run_uuid):
     try: r["inputs"]["Wind"] = meta.WindInputs.dict
     except: pass
 
+    try: r["inputs"]["CoolingLoad"] = meta.CoolingLoadInputs.dict
+    except: pass
+
+    try: r["inputs"]["ExistingChiller"] = meta.ExistingChillerInputs.dict
+    except: pass
+	
     try: r["inputs"]["ExistingBoiler"] = meta.ExistingBoilerInputs.dict
     except: pass
 
     # try: r["inputs"]["Boiler"] = meta.BoilerInputs.dict
     # except: pass
 
+    try: r["inputs"]["HotThermalStorage"] = meta.HotThermalStorageInputs.dict
+    except: pass
+
+    try: r["inputs"]["ColdThermalStorage"] = meta.ColdThermalStorageInputs.dict
+    except: pass
+
     try: r["inputs"]["SpaceHeatingLoad"] = meta.SpaceHeatingLoadInputs.dict
     except: pass
 
     try: r["inputs"]["DomesticHotWaterLoad"] = meta.DomesticHotWaterLoadInputs.dict
+    except: pass
+
+    try: r["inputs"]["CHP"] = meta.CHPInputs.dict
     except: pass
 
     try:
@@ -230,11 +263,24 @@ def results(request, run_uuid):
         except: pass
         try: r["outputs"]["Wind"] = meta.WindOutputs.dict
         except: pass
+        try: r["outputs"]["ExistingChiller"] = meta.ExistingChillerOutputs.dict
+        except: pass
         try: r["outputs"]["ExistingBoiler"] = meta.ExistingBoilerOutputs.dict
         except: pass
         # try: r["outputs"]["Boiler"] = meta.BoilerOutputs.dict
         # except: pass
         try: r["outputs"]["Outages"] = meta.OutageOutputs.dict
+        except: pass
+
+        try: r["outputs"]["HotThermalStorage"] = meta.HotThermalStorageOutputs.dict
+        except: pass
+        try: r["outputs"]["ColdThermalStorage"] = meta.ColdThermalStorageOutputs.dict
+        except: pass
+        try: r["outputs"]["CHP"] = meta.CHPOutputs.dict
+        except: pass
+        try: r["outputs"]["HeatingLoad"] = meta.HeatingLoadOutputs.dict
+        except: pass
+        try: r["outputs"]["CoolingLoad"] = meta.CoolingLoadOutputs.dict
         except: pass
 
         for d in r["outputs"].values():
@@ -256,3 +302,112 @@ def results(request, run_uuid):
             return JsonResponse(resp, status=500)
 
     return JsonResponse(r)
+
+
+def chp_defaults(request):
+    inputs = {
+        "existing_boiler_production_type": request.GET.get("existing_boiler_production_type"),
+        "avg_boiler_fuel_load_mmbtu_per_hour": request.GET.get("avg_boiler_fuel_load_mmbtu_per_hour"),
+        "prime_mover": request.GET.get("prime_mover"),
+        "size_class": request.GET.get("size_class"),
+        "boiler_efficiency": request.GET.get("boiler_efficiency")
+    }
+    try:
+        julia_host = os.environ.get('JULIA_HOST', "julia")
+        http_jl_response = requests.get("http://" + julia_host + ":8081/chp_defaults/", json=inputs)
+        response = JsonResponse(
+            http_jl_response.json()
+        )
+        return response
+
+    except ValueError as e:
+        return JsonResponse({"Error": str(e.args[0])}, status=500)
+
+    except KeyError as e:
+        return JsonResponse({"Error. Missing": str(e.args[0])}, status=500)
+
+    except Exception:
+        exc_type, exc_value, exc_traceback = sys.exc_info()
+        debug_msg = "exc_type: {}; exc_value: {}; exc_traceback: {}".format(exc_type, exc_value.args[0],
+                                                                            tb.format_tb(exc_traceback))
+        log.debug(debug_msg)
+        return JsonResponse({"Error": "Unexpected error in chp_defaults endpoint. Check log for more."}, status=500)
+
+def simulated_load(request):
+    try:
+        valid_keys = ["doe_reference_name","latitude","longitude","load_type","percent_share","annual_kwh",
+                        "monthly_totals_kwh","annual_mmbtu","annual_fraction","annual_tonhour","monthly_tonhour",
+                        "monthly_mmbtu","monthly_fraction","max_thermal_factor_on_peak_load","chiller_cop",
+                        "addressable_load_fraction", "cooling_doe_ref_name", "cooling_pct_share"]
+        for key in request.GET.keys():
+            k = key
+            if "[" in key:
+                k = key.split('[')[0]
+            if k not in valid_keys:
+                raise ValueError("{} is not a valid input parameter".format(key))
+        # Build inputs dictionary to send to http.jl /simulated_load endpoint
+        inputs = {}
+        inputs["latitude"] = float(request.GET['latitude'])  # need float to convert unicode
+        inputs["longitude"] = float(request.GET['longitude'])
+        inputs["load_type"] = request.GET.get('load_type')
+
+        # This parses the GET request way of sending a list/array for doe_reference_name, 
+        # i.e. doe_reference_name[0], doe_reference_name[1], etc
+        if 'doe_reference_name' in request.GET.keys():
+            inputs["doe_reference_name"] = request.GET.get('doe_reference_name')
+        elif 'doe_reference_name[0]' in request.GET.keys():
+            idx = 0
+            doe_reference_name = []
+            percent_share_list = []
+            while 'doe_reference_name[{}]'.format(idx) in request.GET.keys():
+                doe_reference_name.append(request.GET['doe_reference_name[{}]'.format(idx)])
+                if 'percent_share[{}]'.format(idx) in request.GET.keys():
+                    percent_share_list.append(float(request.GET['percent_share[{}]'.format(idx)]))
+                idx += 1
+            inputs["doe_reference_name"] = doe_reference_name
+            inputs["percent_share_list"] = percent_share_list
+
+        # When wanting cooling profile based on building type(s) for cooling, need separate cooling building(s)
+        if 'cooling_doe_ref_name' in request.GET.keys():
+            inputs["cooling_doe_ref_name"] = request.GET.get('cooling_doe_ref_name')
+        elif 'cooling_doe_ref_name[0]' in request.GET.keys():
+            idx = 0
+            cooling_doe_ref_name = []
+            cooling_pct_share_list = []
+            while 'cooling_doe_ref_name[{}]'.format(idx) in request.GET.keys():
+                cooling_doe_ref_name.append(request.GET['cooling_doe_ref_name[{}]'.format(idx)])
+                if 'cooling_pct_share[{}]'.format(idx) in request.GET.keys():
+                    cooling_pct_share_list.append(float(request.GET['cooling_pct_share[{}]'.format(idx)]))
+                idx += 1
+            inputs["cooling_doe_ref_name"] = cooling_doe_ref_name
+            inputs["cooling_pct_share"] = cooling_pct_share_list                
+
+        # Build the rest of inputs for http.jl /simulated_load endpoint
+        other_keys_types = ["annual", "monthly", "max_thermal_factor", "chiller_cop", "addressable"]
+        for key in valid_keys:
+            for key_type in other_keys_types:
+                if key_type in key:
+                    value = request.GET.get(key)
+                    if value is not None:
+                        inputs[key] = value
+
+        julia_host = os.environ.get('JULIA_HOST', "julia")
+        http_jl_response = requests.get("http://" + julia_host + ":8081/simulated_load/", json=inputs)
+        response = JsonResponse(
+            http_jl_response.json()
+        )
+        
+        return response
+
+    except ValueError as e:
+        return JsonResponse({"Error": str(e.args[0])}, status=500)
+
+    except KeyError as e:
+        return JsonResponse({"Error. Missing": str(e.args[0])}, status=500)
+
+    except Exception:
+        exc_type, exc_value, exc_traceback = sys.exc_info()
+        debug_msg = "exc_type: {}; exc_value: {}; exc_traceback: {}".format(exc_type, exc_value.args[0],
+                                                                            tb.format_tb(exc_traceback))
+        log.debug(debug_msg)
+        return JsonResponse({"Error": "Unexpected error in simulated_load endpoint. Check log for more."}, status=500)